# contracts/openapi.json
{
  "components": {
    "responses": {
      "ErrorResponse": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Stable machine-readable error envelope."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error_code": {
            "type": "string"
          },
          "field_errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "error_code",
          "message"
        ],
        "type": "object"
      },
...
      "FieldError": {
        "additionalProperties": false,
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ],
        "type": "object"
      },
...
      "GroupResponse": {
        "additionalProperties": false,
        "description": "Server-owned group identity and settlement policy.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "owner_account_id": {
            "title": "Owner Account Id",
            "type": "string"
          },
          "settlementPolicy": {
            "enum": [
              "owner_only",
              "any_member"
            ],
            "title": "Settlementpolicy",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "owner_account_id",
          "settlementPolicy"
        ],
        "title": "GroupResponse",
        "type": "object"
      },
...
      "GroupUpdateRequest": {
        "additionalProperties": false,
        "description": "The only mutable group setting exposed by the MVP.",
        "properties": {
          "settlementPolicy": {
            "enum": [
              "owner_only",
              "any_member"
            ],
            "title": "Settlementpolicy",
            "type": "string"
          }
        },
        "required": [
          "settlementPolicy"
        ],
        "title": "GroupUpdateRequest",
        "type": "object"
      },
...
      "OutingResponse": {
        "additionalProperties": false,
        "description": "Server-owned active/archived outing history shape.",
        "properties": {
          "archived": {
            "title": "Archived",
            "type": "boolean"
          },
          "archived_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "group_id": {
            "title": "Group Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "id",
          "group_id",
          "name",
          "archived"
        ],
        "title": "OutingResponse",
        "type": "object"
      },
      "OutingWriteRequest": {
        "additionalProperties": false,
        "description": "The only editable outing field in this bounded lifecycle slice.",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "OutingWriteRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "cc_session": {
        "description": "Opaque server-recognized session cookie.",
        "in": "cookie",
        "name": "cc_session",
        "type": "apiKey"
      },
      "csrf_token": {
        "description": "Required on unsafe cookie-authenticated operations.",
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Cuentas Claras API",
    "version": "0.1.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/groups/{group_id}": {
      "get": {
        "description": "Return the authenticated group's server-owned settings.",
        "operationId": "get_group_api_v1_groups__group_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "409": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "422": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "security": [
          {
            "cc_session": []
          }
        ],
        "summary": "Get Group",
        "tags": [
          "groups"
        ]
      },
      "patch": {
        "description": "Update only settlement policy; authorization remains in GroupService.",
        "operationId": "update_group_api_v1_groups__group_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "string"
            }
          },
          {
            "description": "Must match the readable cc_csrf cookie.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "409": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "422": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "security": [
          {
            "cc_session": []
          }
        ],
        "summary": "Update Group",
        "tags": [
          "groups"
        ]
      }
    },
...
    "/api/v1/groups/{group_id}/outings": {
      "get": {
        "description": "List active and archived outings in stable creation order.",
        "operationId": "list_outings_api_v1_groups__group_id__outings_get",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OutingResponse"
                  },
                  "title": "Response List Outings Api V1 Groups  Group Id  Outings Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "409": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "422": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "security": [
          {
            "cc_session": []
          }
        ],
        "summary": "List Outings",
        "tags": [
          "outings"
        ]
      },
      "post": {
        "description": "Create an active outing for any active group member.",
        "operationId": "create_outing_api_v1_groups__group_id__outings_post",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "string"
            }
          },
          {
            "description": "Must match the readable cc_csrf cookie.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OutingWriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "409": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "422": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "security": [
          {
            "cc_session": []
          }
        ],
        "summary": "Create Outing",
        "tags": [
          "outings"
        ]
      }
    },
    "/api/v1/groups/{group_id}/outings/{outing_id}": {
      "delete": {
        "description": "Delete an empty active outing as its server-authorized owner.",
        "operationId": "delete_outing_api_v1_groups__group_id__outings__outing_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "outing_id",
            "required": true,
            "schema": {
              "title": "Outing Id",
              "type": "string"
            }
          },
          {
            "description": "Must match the readable cc_csrf cookie.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "409": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "422": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "security": [
          {
            "cc_session": []
          }
        ],
        "summary": "Delete Outing",
        "tags": [
          "outings"
        ]
      },
      "get": {
        "description": "Read an active or archived outing in the requested group.",
        "operationId": "get_outing_api_v1_groups__group_id__outings__outing_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "outing_id",
            "required": true,
            "schema": {
              "title": "Outing Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "409": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "422": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "security": [
          {
            "cc_session": []
          }
        ],
        "summary": "Get Outing",
        "tags": [
          "outings"
        ]
      },
      "patch": {
        "description": "Edit only the name of an active outing.",
        "operationId": "edit_outing_api_v1_groups__group_id__outings__outing_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "outing_id",
            "required": true,
            "schema": {
              "title": "Outing Id",
              "type": "string"
            }
          },
          {
            "description": "Must match the readable cc_csrf cookie.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OutingWriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "409": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "422": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "security": [
          {
            "cc_session": []
          }
        ],
        "summary": "Edit Outing",
        "tags": [
          "outings"
        ]
      }
    },
    "/api/v1/groups/{group_id}/outings/{outing_id}/archive": {
      "post": {
        "description": "Archive an outing as a server-authorized owner operation.",
        "operationId": "archive_outing_api_v1_groups__group_id__outings__outing_id__archive_post",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "outing_id",
            "required": true,
            "schema": {
              "title": "Outing Id",
              "type": "string"
            }
          },
          {
            "description": "Must match the readable cc_csrf cookie.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "409": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "422": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "security": [
          {
            "cc_session": []
          }
        ],
        "summary": "Archive Outing",
        "tags": [
          "outings"
        ]
      }
    },
    "/api/v1/groups/{group_id}/outings/{outing_id}/unarchive": {
      "post": {
        "description": "Restore an archived outing to active state as its owner.",
        "operationId": "unarchive_outing_api_v1_groups__group_id__outings__outing_id__unarchive_post",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "outing_id",
            "required": true,
            "schema": {
              "title": "Outing Id",
              "type": "string"
            }
          },
          {
            "description": "Must match the readable cc_csrf cookie.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "409": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "422": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "security": [
          {
            "cc_session": []
          }
        ],
        "summary": "Unarchive Outing",
        "tags": [
          "outings"
        ]
      }
    },
...
  }
}